requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
OdooRPC==0.10.1
sqlparse==0.5.5
